+++ src/client/src/components/Login/steps/Form/FormError.js
@@ -1,11 +1,11 @@
-import { faExclamationTriangle } from "@fortawesome/free-solid-svg-icons";
+import { faTriangleExclamation } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import PropTypes from "prop-types";
 import React from "react";
 
 const FormError = ({ errorMessage }) => (
   <div className="login__notif login__notif--error shake-horizontal">
-    <FontAwesomeIcon icon={faExclamationTriangle} />
+    <FontAwesomeIcon icon={faTriangleExclamation} />
     <p>{errorMessage}</p>
   </div>
 );